refactor: Relocate use case modules to domain layer
This commit moves the use case modules from the `data/usecase` directory to a new `domain` directory. This change aligns the project structure with common architectural patterns where domain logic, including use cases, resides in a separate domain layer.

- Moved all use case classes (e.g., `GetArticlesUseCase`, `GetArticleByIdUseCase`) from `data.usecase` package to `com.kesicollection.domain` package.
- Updated all import statements across feature modules (`article`, `articles`, `discover`, `audioplayer`) to reflect the new package structure for use cases.
- Renamed the Gradle module `data:usecase` to `domain`.
- Updated `settings.gradle.kts` to include the new `:domain` module and removed the old `:data:usecase` module.
- Updated `build-logic/convention/src/main/kotlin/AndroidFeatureConventionPlugin.kt` to depend on `project(":domain")` instead of `project(":data:usecase")`.
- Deleted the old `data/usecase` directory and its contents.
- Created a new `domain` directory with the relocated use case files and a new `build.gradle.kts` and `.gitignore` file.
