+++ build-logic/convention/src/main/kotlin/AndroidFeatureConventionPlugin.kt
@@ -51,7 +51,7 @@ class AndroidFeatureConventionPlugin : Plugin<Project> {
 
             dependencies {
                 implementation(project(":core:uisystem"))
-                implementation(project(":data:usecase"))
+                implementation(project(":domain"))
                 implementation(libs.findLibrary("androidx.hilt.navigation.compose").get())
                 implementation(libs.findLibrary("androidx.lifecycle.runtimeCompose").get())
                 implementation(libs.findLibrary("androidx.lifecycle.viewModelCompose").get())
