+++ feature/article/src/test/kotlin/com/kesicollection/feature/article/intent/FetchArticleIntentProcessorTest.kt
@@ -1,7 +1,7 @@
 package com.kesicollection.feature.article.intent
 
 import com.kesicollection.core.model.ErrorState
-import com.kesicollection.data.usecase.GetArticleByIdUseCase
+import com.kesicollection.domain.GetArticleByIdUseCase
 import com.kesicollection.feature.article.ArticleErrors
 import com.kesicollection.feature.article.UiArticleState
 import com.kesicollection.feature.article.components.Code
