+++ feature/discover/src/main/kotlin/com/kesicollection/feature/discover/intentprocessor/FetchFeatureItemsIntentProcessor.kt
@@ -4,18 +4,16 @@ import com.kesicollection.core.app.CrashlyticsWrapper
 import com.kesicollection.core.app.IntentProcessor
 import com.kesicollection.core.app.Reducer
 import com.kesicollection.core.model.ErrorState
-import com.kesicollection.data.usecase.GetDiscoverContentUseCase
+import com.kesicollection.domain.GetDiscoverContentUseCase
 import com.kesicollection.feature.discover.DiscoverErrors
 import com.kesicollection.feature.discover.UiState
 import com.kesicollection.feature.discover.asDiscoverContent
-import com.kesicollection.feature.discover.contentSample
-import kotlinx.coroutines.delay
 import java.util.concurrent.CancellationException
 
 /**
  * Processes the intent to fetch feature items for the Discover screen.
  *
- * This processor utilizes the [GetDiscoverContentUseCase] to retrieve discover content.
+ * This processor utilizes the [com.kesicollection.domain.GetDiscoverContentUseCase] to retrieve discover content.
  * On success, it maps the result to [UiState.DiscoverContent] and updates the UI through the provided [Reducer].
  * On failure, it maps the exception to [UiState.Error] with a [DiscoverErrors.GenericError]
  * and updates the UI.
