+++ settings.gradle.kts
@@ -36,3 +36,4 @@ include(":data:usecase")
 include(":data:datastore")
 include(":core:app")
 include(":feature:discover")
+include(":domain")
